Add more info popover to createperview and editperview
