+++ Ashish_P/perview/src/styles/stylesheets/PerviewLayouts/PerviewCard/pervieweditmodal.scss
@@ -94,7 +94,7 @@ $lines-to-show: 4;
   }
 
   &product-edits {
-    align-items: flex-end;
+    align-items: flex-start;
     margin-top: 30px;
   }
