+++ Ashish_P/perview/src/styles/stylesheets/SharedComponents/pricepopover.scss
@@ -1,0 +1,5 @@
+.price__popover {
+  font-size: 12px;
+  font-family: roboto;
+  font-weight: lighter;
+}
