+++ Ashish_P/perview/src/styles/stylesheets/createperviewmodal.scss
@@ -136,7 +136,7 @@ $lines-to-show: 4;
   }
 
   &product-details {
-    align-items: flex-end;
+    align-items: flex-start;
     margin-top: 30px;
   }
